fix 16 byte running times

+++ service/server/src/main/java/com/hansun/server/commu/LinkManger.java
@@ -99,19 +99,21 @@ public class LinkManger {
         //需要判断当前设备所有端口是否还有订单，和设备运行时间是否相符合
         List<Device> deviceList = deviceService.getDevicesByDeviceBox(deviceName);
 
-        for (Device device : deviceList) {
-            Order order = orderService.getOrder(device.getId());
-            MsgTime msgTime = (MsgTime) portMap.get(device.getPort());
+        if (deviceList != null) {
+            for (Device device : deviceList) {
+                Order order = orderService.getOrder(device.getId());
+                MsgTime msgTime = (MsgTime) portMap.get(device.getPort());
 
-            if (order != null && msgTime.getTime() == 0) {
-                //如果订单还在缓存中，但是结束时在当前时间之前，则需要从缓存中删除该订单
-                if (Instant.now().isAfter(order.getCreateTime().plus(Duration.ofMinutes(order.getDuration())))
-                        || Instant.now().isAfter(order.getStartTime().plus(Duration.ofMinutes(order.getDuration())))) {
-                    //设备没有收到后续结束报文，所以收到心跳消息，判断当前设备是否还在运行，如果指示时间为0，而订单是运行中，则更新订单为finish
-                    if (order.getOrderStatus() == OrderStatus.SERVICE) {
-                        orderService.deleteOrder(device.getId());
-                    } else {
-                        orderService.removeOrder(device.getId());
+                if (order != null && msgTime.getTime() == 0) {
+                    //如果订单还在缓存中，但是结束时在当前时间之前，则需要从缓存中删除该订单
+                    if (Instant.now().isAfter(order.getCreateTime().plus(Duration.ofMinutes(order.getDuration())))
+                            || Instant.now().isAfter(order.getStartTime().plus(Duration.ofMinutes(order.getDuration())))) {
+                        //设备没有收到后续结束报文，所以收到心跳消息，判断当前设备是否还在运行，如果指示时间为0，而订单是运行中，则更新订单为finish
+                        if (order.getOrderStatus() == OrderStatus.SERVICE) {
+                            orderService.deleteOrder(device.getId());
+                        } else {
+                            orderService.removeOrder(device.getId());
+                        }
                     }
                 }
             }
